// _backup_before_cleanup_20251126_012011/src/types/database.ts
export type Tables<T extends keyof Database['public']['Tables']> = Database['public']['Tables'][T]['Row']
// ...
export type LetterCertificate = Tables<'letters_certificates'>
// ...
export type LetterCertificateInsert = Database['public']['Tables']['letters_certificates']['Insert']
// ...
export type LetterCertificateType = 'letter' | 'certificate' | 'award' | 'diploma'